% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_whitespaces_to_seperate_chunks}
\alias{add_whitespaces_to_seperate_chunks}
\title{Add whitespace after single tex chunk.}
\usage{
add_whitespaces_to_seperate_chunks(x)
}
\arguments{
\item{x}{texcode object.}
}
\value{
A textcode object with whitespace added where needed.
}
\description{
The LaTeX commands need to be seperated by whitespaces or linebreaks.
But after the $ evironment delimiter they are not allowed.
The function:
- adds a whitespace at the end of every chunk except
 after chunks beginning or ending with $
- adds no whitespace at the end of chunks followed by a chunk onyl
containing $
}
\details{
A cleaner approach would use objects that come along with the
space to insert but that would mean reprogramming everything.
}
\examples{
x <- "$" + lex(10) + lex("a") + "$\\\\n"
}
\keyword{internal}

