% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.texcode}
\alias{as.texcode}
\title{Convert to \code{texcode} object.}
\usage{
as.texcode(x)
}
\arguments{
\item{x}{Some R object.}
}
\description{
Convert to \code{texcode} object.
}
\details{
Currently only character vectors can be converted.
}
\keyword{internal}

