% Generated by roxygen2 (4.0.2): do not edit by hand
\name{\%_\%}
\alias{\%_0\%}
\alias{\%_1\%}
\alias{\%_2\%}
\alias{\%_3\%}
\alias{\%_4\%}
\alias{\%_5\%}
\alias{\%_\%}
\alias{\%__\%}
\title{Operators to allow for easy combination of texcode objects}
\usage{
x \%_\% y

x \%__\% y

x \%_0\% y

x \%_1\% y

x \%_2\% y

x \%_3\% y

x \%_4\% y

x \%_5\% y
}
\arguments{
\item{x,y}{\code{texcode} object, string or matrix.}
}
\description{
Operators to allow for easy combination of texcode objects
}
\keyword{internal}

