% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcode.r
\name{is.texcode}
\alias{is.texcode}
\title{Check if object is a \code{texcode} object.}
\usage{
is.texcode(x)
}
\arguments{
\item{x}{R object.}
}
\description{
Check if object is a \code{texcode} object.
}
\keyword{internal}
