% Generated by roxygen2 (4.0.2): do not edit by hand
\name{latexcode_bordermatrix}
\alias{latexcode_bordermatrix}
\title{Create LaTeX code for bordermatrix.}
\usage{
latexcode_bordermatrix(x, corner = "")
}
\arguments{
\item{x}{A matrix object with column and/or rownames.}

\item{corner}{Entry for upper left corner, usually empty.}
}
\value{
List with LaTeX code fragments.
}
\description{
Create LaTeX code for bordermatrix.
}
\keyword{internal}

