% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lex}
\alias{lex}
\title{Inject LaTeX code.
Converts LaTeX code into texcode object.}
\usage{
lex(x)
}
\arguments{
\item{x}{Some arbitrary LaTeX code.}
}
\value{
Object of class texcode.
}
\description{
Inject LaTeX code.
Converts LaTeX code into texcode object.
}
\author{
Mark Heckmann
}
\keyword{internal}

