% Generated by roxygen2 (4.0.1): do not edit by hand
\name{mat2tex_options}
\alias{mat2tex_options}
\title{get and set mat2tex default options}
\usage{
mat2tex_options(...)
}
\arguments{
\item{...}{Enter the options you want to change as \code{name = value}
pairs separated by commas.}
}
\description{
get and set mat2tex default options
}
\examples{
\dontrun{

 # show all mat2tex options
 mat2tex_options()

 # change default for number of digits to round to
 mat2tex_options(digits=3)

 # change default for matrix type and digits
 mat2tex_options(digits=3, mtype="bmatrix")

 # change default tranpose sign (e.g. used in xmt)
 mat2tex_options(transpose.sign="^\\\\intercal")

}
}

