% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcode.r
\name{matrix_name}
\alias{matrix_name}
\title{Generate the assigned variable in standard matrix notation as latex}
\usage{
matrix_name(mat = NULL, var = "X", m = NULL, n = NULL)
}
\arguments{
\item{mat}{matrix object}

\item{var}{string with the name of the variable}

\item{m}{user defined number of rows}

\item{n}{user defined number of columns}
}
\description{
Generate the assigned variable in standard matrix notation as latex
}
