% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Ops.texcode}
\alias{Ops.texcode}
\title{Overloading "+" and "\%\%" operator to allow for easy combination of texcode objects.}
\usage{
\method{Ops}{texcode}(e1, e2)
}
\arguments{
\item{e1}{texcode object}

\item{e2}{texcode object}
}
\description{
Overloading "+" and "\%\%" operator to allow for easy combination of texcode objects.
}
\keyword{internal}

