% Generated by roxygen2 (4.0.2): do not edit by hand
\name{s}
\alias{s}
\title{Insert horizontal spaces in formula}
\usage{
s(em = 0)
}
\arguments{
\item{em}{Numeric. Width of inserted space in multiples
of letter \code{M}. The default is \code{0}.
Negative spaces are allowed.}
}
\description{
The function is a wrapper around the
LaTeX \code{\\mkern} command. It will produce
horizintal spaces with a width given as multiples of the
letter \code{m}. E.g. \code{s(2)} equates the LaTeX code
\code{\\mkern2em}.
}
\examples{
"$ 1 + 1 =" \%_\% s(2) \%_\% "2$"
 xx("1+1=", s(2), "2")
}

