% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xb}
\alias{xb}
\alias{xe}
\title{Begin or end LaTeX math environment.}
\usage{
xb(e = NA, label = NULL)

xe(e = 1)
}
\arguments{
\item{e}{Numeric indicating the type of math environment to begin or end
(\code{1 = equation, 2= equation*, 3=$, 4=align, 5=align*}).
Default is \code{1 = 'equation'}.}

\item{label}{Optional LaTeX equation label}
}
\value{
Object of class texcode.
}
\description{
\code{xb} will begin, \code{xb} will end a math environment.


}
\author{
Mark Heckmann
}

