% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xx}
\alias{xx}
\title{Concatenate \code{texcode} chunks.}
\usage{
xx(..., e = 1, label = NULL)
}
\arguments{
\item{...}{\code{texcode} objects or objects that can ce converted
into texcode objects.}
}
\value{
Object of class \code{texcode}.
}
\description{
Concatenate \code{texcode} chunks.
}
\author{
Mark Heckmann
}

