% Generated by roxygen2 (4.0.1): do not edit by hand
\name{\%_\%}
\alias{\%_\%}
\alias{\%__\%}
\title{"\%\_\%" operator to allow for easy combination of texcode objects}
\usage{
x \%_\% y

x \%__\% y
}
\arguments{
\item{x,y}{\code{texcode} object, string or matrix.}
}
\description{
"\%\_\%" operator to allow for easy combination of texcode objects
}
\keyword{internal}

