% Generated by roxygen2 (4.0.2): do not edit by hand
\name{latexcode_matrix}
\alias{latexcode_matrix}
\title{Create LaTeX code for several matrix types defined in the amsmath package.}
\usage{
latexcode_matrix(x, mtype)
}
\arguments{
\item{x}{A matrix object.}

\item{mtype}{LaTeX matrix type.}
}
\value{
List with LaTeX code fragments.
}
\description{
Create LaTeX code for several matrix types defined in the amsmath package.
}
\keyword{internal}

