% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat2tex.R
\docType{package}
\name{mat2tex-package}
\alias{mat2tex-package}
\alias{mat2tex}
\title{Tools to combine R and LaTeX code for use with knitr or sweave}
\description{
Easily create LaTeX equations with R matrices for .Rnw files
}
\references{
For an introduction visit \url{http://markheckmann.github.io/mat2tex/}.
}
\author{
Mark Heckmann \email{heckmann@uni-bremen.de}
}
\keyword{package}
