% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{mat2tex-package}
\alias{mat2tex}
\alias{mat2tex-package}
\title{Tools to combine R and LaTeX code for use with knitr or sweave}
\description{
Easily create LaTeX equations with R matrices for .Rnw files
}
\author{
Mark Heckmann \email{heckmann@uni-bremen.de}
}
\references{
For an introduction visit \url{http://markheckmann.github.io/mat2tex/}.
}
\keyword{package}

