% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xb}
\alias{xb}
\alias{xe}
\title{Begin or end LaTeX math environment.}
\usage{
xb(e = NA, label = NULL)

xe(e = 1)
}
\arguments{
\item{e}{Math environment to use (numeric or string).
If \code{NA} the default
is used as defined in \code{mat2tex_options()$mathenv}.
If set to \code{NULL}, no environment is added, just the plain
math code is returned. See details.}

\item{label}{Optional LaTeX equation label}
}
\value{
Object of class texcode.
}
\description{
\code{xb} will begin, \code{xb} will end a math environment.


}
\details{
The available math environments are
          \code{1=$$, 2=$, 3=equation, 4=equation*, 5=align, 6=align*,
         7=gather, 8=gather*, 9=multiline, 10=multiline*, 11=split}.
         You can either supply the corresponding numeric or the name of
         the math environment in argument \code{e}.
         For \code{Rmd} file only math environment (argument \code{e}) 1 and 2 are relevant.
         When using \code{Rnw} files make sure to include the \code{amsmath} package
         the document preamble as most environments are defined in it.
}
\author{
Mark Heckmann
}

