% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texcode.r
\name{xc}
\alias{xc}
\title{Inject LaTeX code.
Converts LaTeX code into texcode object.}
\usage{
xc(x)
}
\arguments{
\item{x}{Some arbitrary LaTeX code.}
}
\value{
Object of class texcode.
}
\description{
Inject LaTeX code.
Converts LaTeX code into texcode object.
}
\author{
Mark Heckmann
}
