% Generated by roxygen2 (4.0.1): do not edit by hand
\name{xm}
\alias{xm}
\alias{xmt}
\title{Convert a matrix to LaTeX code.}
\usage{
xm(x, digits = NA, mtype = NA, round = NA)

xmt(x, digits = 2, mtype = "pmatrix", round = TRUE)
}
\arguments{
\item{x}{A matrix object.}

\item{mtype}{LaTeX matrix type, i.e. round braces, curly braces etc.
Available types are \code{"matrix", "pmatrix", "bmatrix",
"Bmatrix", "vmatrix", "Vmatrix"}.}

\item{digits}{Number of digits to display. Only trimmin, no rounding.}
}
\value{
Object of class texcode.
}
\description{
Convert a matrix to LaTeX code.
}
\examples{
m <- matrix(1:9, ncol=3)
xm(m)

# change digits

"$" \%\% xm(m) \%\% "$"

"$" + xm(m) + "$"
}
\author{
Mark Heckmann
}

